Add Copy Tool
Added ability for creative-players (or non-creative according to config) to use an item with specific NBT tag on a tent door. Doing so gives the player the tent item without deconstructing the tent. Useful for admins to go into tent dimension and re-create lost tent items without having to do a lot of math.