It's Alive!
Somehow, the system I had in mind is actually working! Replaced all uses of StructureType and its subclasses with the wrapper class StructureData. This class holds current and previous values for Tent Type, Width, and Depth. It currently does not store tent X/Z location, but that might be ported over soon. All tent data is stored in NBT, even tent type and size. Refined recipes to allow for this. Added basic depth upgrade recipe (in-progress). Larger changes to StructureBase - I'm hoping to get it to the point where its child classes are used as static singletons, but I'd have to re-think a lot of code. Mostly where to store the Blueprints.