JEI kinda works
JEI recipes now show more or less correctly, with the exception of depth recipes only showing the highest upgrade obtainable instead and not the intermediate steps required. Close enough.
This required more changes to the recipe system, but in case anyone has their own recipe handler, I added a check for a "disabled" tag in the recipe that will stop the recipe from loading.